+++ src/Illuminate/Foundation/Testing/Concerns/MakesHttpRequests.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Illuminate\Foundation\Testing;
+namespace Illuminate\Foundation\Testing\Concerns;
 
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
